+++ src/main/java/ctoutweb/lalamiam/model/CommandInformationToUpdate.java
@@ -1,6 +1,5 @@
 package ctoutweb.lalamiam.model;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -16,8 +15,8 @@ import java.util.List;
  * @param slotTime
  */
 public record CommandInformationToUpdate(
-        BigInteger storeId,
-        BigInteger commandId,
+        Long storeId,
+        Long commandId,
         String clientPhone,
         List<ProductWithQuantity> selectProducts,
         LocalDateTime slotTime,
