+++ src/main/java/ctoutweb/lalamiam/model/dto/AddProductResponseDto.java
@@ -1,10 +1,9 @@
 package ctoutweb.lalamiam.model.dto;
 
-import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 public record AddProductResponseDto(
-        BigInteger id,
+        Long id,
         String name,
         Double price,
         String description,
