+++ src/main/java/ctoutweb/lalamiam/model/dto/UpdateProductDto.java
@@ -1,13 +1,11 @@
 package ctoutweb.lalamiam.model.dto;
 
-import java.math.BigInteger;
-
 public record UpdateProductDto(
-        BigInteger productId,
+        Long productId,
         String name,
         Double price,
         String description,
         Integer preparationTime,
         String photo,
-        BigInteger storeId) {
+        Long storeId) {
 }
