+++ src/main/java/ctoutweb/lalamiam/repository/StoreRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.math.BigInteger;
 
 @Repository
-public interface StoreRepository extends JpaRepository<StoreEntity, BigInteger> {
+public interface StoreRepository extends JpaRepository<StoreEntity, Long> {
   @Query(value = "select count(*) from sc_lalamiam.store", nativeQuery = true)
   public long countAll();
 }
